refactor: 인물 도메인 모델 변경으로 인한 홈 화면 수정 (#27)
- 이미지 업로드 아이콘 변경 포함

+++ feature/home/src/main/java/com/example/metasearch/feature/home/HomeUiState.kt
@@ -30,7 +30,7 @@ sealed interface HomeUiEvent : CircuitUiEvent {
      * 상단 인물 클릭
      */
     data class OnPersonClick(
-        val personId: Int,
+        val personId: Long,
     ) : HomeUiEvent
 
     /**
